requirements.txt
chromadb==0.5.4
langchain==0.2.10
langchain-community==0.2.9
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langchainhub==0.1.20
openai==1.36.1
pypdf==4.3.0
python-dotenv==1.0.1
streamlit==1.36.0
tiktoken==0.7.0